requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cmdstanpy==1.2.5
comm==0.2.3
contourpy==1.3.2
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docling==2.55.0
docling-core==2.48.4
docling-ibm-models==3.9.1
docling-parse==4.5.0
easyocr==1.7.2
ecdsa==0.19.1
einops==0.8.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.12.0
Faker==37.8.0
fastapi==0.104.1
fastjsonschema==2.21.2
feedparser==6.0.12
filelock==3.19.1
filetype==1.2.0
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.3
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gspread==6.2.1
h11==0.16.0
hf-xet==1.1.10
holidays==0.82
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
json5==0.12.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaggle==1.7.4.5
kiwisolver==1.4.8
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-huggingface==0.0.3
langchain-mistralai==0.1.13
langchain-ollama==0.1.0
langchain-text-splitters==0.2.4
langsmith==0.1.147
lark==1.3.0
latex2mathml==3.78.1
lazy_loader==0.4
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.9.2
mistune==3.1.4
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
narwhals==1.47.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.1
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.0
openai==2.7.1
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.2.0
pluggy==1.6.0
polyfactory==2.22.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
prophet==1.1.7
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
pylatexenc==2.10
PyMySQL==1.1.0
pyodbc==5.0.1
pyparsing==3.2.3
pypdf==6.1.1
pypdfium2==4.30.0
pyspark==4.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.3.0
python-json-logger==4.0.0
python-multipart==0.0.6
python-pptx==1.0.2
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semchunk==2.2.2
Send2Trash==1.8.3
sentence-transformers==5.1.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.4.4
stack-data==0.6.3
stanio==0.5.1
starlette==0.27.0
statsmodels==0.14.5
streamlit==1.47.0
stringzilla==4.2.1
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
torch==2.8.0
torch_geometric==2.4.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typer==0.19.2
types-python-dateutil==2.9.0.20251008
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.14
xlsxwriter==3.2.9
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
